build(generate): main readme gen refactor
