+++ scripts/generate/main.ts
@@ -1,20 +1,12 @@
-#!/usr/bin/env -S deno run --allow-env --allow-read --allow-write --allow-net
-
+#!/usr/bin/env -S deno run -A
 import { join } from "std/path/mod.ts";
-import Handlebars from "handlebars";
 
 import { portsSchema, REPO_ROOT, schema } from "@/deps.ts";
 import { PortsSchema, UserStylesSchema } from "@/types/mod.d.ts";
 import { syncIssueLabels } from "./labels.ts";
-import { updateFile, updateReadme, validateYaml } from "./utils.ts";
+import { updateFile, validateYaml } from "./utils.ts";
+import { generateMainReadme } from "./readme-repo.ts";
 import { generateStyleReadmes } from "./readme-styles.ts";
-
-type Metadata = {
-  userstyles: UserStylesSchema.Userstyles;
-};
-type PortMetadata = { categories: PortsSchema.Categories };
-
-export type MappedPort = UserStylesSchema.Userstyle & { path: string };
 
 const userstylesYaml = Deno.readTextFileSync(
   join(REPO_ROOT, "scripts/userstyles.yml"),
@@ -24,8 +16,8 @@ const portsYaml = await fetch(
 ).then((res) => res.text());
 
 const [portsData, userstylesData] = await Promise.all([
-  await validateYaml<PortMetadata>(portsYaml, portsSchema),
-  await validateYaml<Metadata>(userstylesYaml, schema),
+  await validateYaml<PortsSchema.PortsSchema>(portsYaml, portsSchema),
+  await validateYaml<UserStylesSchema.UserstylesSchema>(userstylesYaml, schema),
 ]).catch((e) => {
   console.error(e);
   Deno.exit(1);
@@ -36,71 +28,34 @@ if (!userstylesData.userstyles) {
   Deno.exit(1);
 }
 
-const categorized = Object.entries(userstylesData.userstyles)
-  .reduce((acc, [slug, { category, ...port }]) => {
-    // initialize category array if it doesn't exist
-    acc[category] ??= [];
-
-    acc[category].push({ path: `styles/${slug}`, category, ...port });
-
-    // Sort by name, first array entry if necessary
-    acc[category].sort((a, b) =>
-      [a.name].flat()[0].localeCompare([b.name].flat()[0])
-    );
-    return acc;
-  }, {} as Record<string, MappedPort[]>);
-
-const portListData = portsData.categories
-  .filter((category) => categorized[category.key] !== undefined)
-  .map((category) => {
-    return { meta: category, ports: categorized[category.key] };
-  });
-
-const portContent = Handlebars.compile(`{{#each category}}
-<details open>
-<summary>{{emoji}} {{name}}</summary>
-
-{{#each ports}}
-- [{{#each name}}{{ this }}{{#unless @last}}, {{/unless}}{{/each}}]({{ path }})
-{{/each}}
-
-</details>
-{{/each}}`)({
-  category: portListData.map(({ meta, ports }) => {
-    return {
-      emoji: meta.emoji,
-      name: meta.name,
-      ports: ports.map(({ name, path }) => {
-        return {
-          name: [name].flat(),
-          path,
-        };
-      }),
-    };
-  }),
-});
-
-const readmePath = join(REPO_ROOT, "README.md");
-await updateFile(
-  readmePath,
-  updateReadme({
-    readme: Deno.readTextFileSync(readmePath),
-    section: "userstyles",
-    newContent: portContent,
-  }),
-  false,
-).catch((e) => console.error(e));
-
-syncIssueLabels(userstylesData.userstyles);
+/**
+ * Generate the main README.md, listing all ports as a table of contents
+ */
+await generateMainReadme(userstylesData.userstyles, portsData);
+/**
+ * Generate README.md files for each style
+ */
 generateStyleReadmes(userstylesData.userstyles);
+/**
+ * Keep
+ * - `.github/issue-labeler.yml`
+ * - `.github/labels.yml`
+ * - `.github/pr-labeler.yml`
+ * in sync with the userstyle metadata.
+ */
+syncIssueLabels(userstylesData.userstyles);
 
-const ownersPath = join(REPO_ROOT, ".github/CODEOWNERS");
-const ownersContent = Object.entries(userstylesData.userstyles)
-  .map(([key, style]) => {
-    const currentMaintainers = style.readme["current-maintainers"]
-      .map((maintainer) => `@${maintainer.url.split("/").pop()}`)
-      .join(" ");
-    return `/styles/${key} ${currentMaintainers}`;
-  })
-  .join("\n");
-updateFile(ownersPath, ownersContent);
+/**
+ * Keep `.github/CODEOWNERS` in sync with the userstyle metadata.
+ */
+await updateFile(
+  join(REPO_ROOT, ".github/CODEOWNERS"),
+  Object.entries(userstylesData.userstyles)
+    .map(([slug, { readme }]) => {
+      const currentMaintainers = readme["current-maintainers"]
+        .map((maintainer) => `@${maintainer.url.split("/").pop()}`)
+        .join(" ");
+      return `/styles/${slug} ${currentMaintainers}`;
+    })
+    .join("\n"),
+);
